+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-09-01 16:49:51
-// Commit: style(test): harmonize light and dark theme palette for test
+// Updated at: 2026-09-01 18:34:12
+// Commit: test(test): implement widget integration tests for test
 
 class testComponentState {
-  final String commitRef = '4118';
-  final String lastMessage = 'style(test): harmonize light and dark theme palette for test';
-  final DateTime timestamp = DateTime.parse('2026-09-01T16:49:51+05:00');
+  final String commitRef = '4120';
+  final String lastMessage = 'test(test): implement widget integration tests for test';
+  final DateTime timestamp = DateTime.parse('2026-09-01T18:34:12+05:00');
 }